+++ src/components/Field/FieldTypes.tsx
@@ -1,7 +1,7 @@
 import { FieldDefAddress } from "@root/components/Field/FormFieldAddress";
 import { FieldDefAdvancedSelection } from "@root/components/Field/FormFieldAdvancedSelection";
-import { FieldDefCheckbox } from "@root/forms/FormFieldCheckbox";
-import { FieldDefChip } from "@root/forms/FormFieldChipSingleSelect";
+import { FieldDefCheckbox } from "@root/components/Field/FormFieldCheckbox";
+import { FieldDefChip } from "@root/components/Field/FormFieldChipSingleSelect";
 import { FieldDefColor } from "@root/forms/FormFieldColorPicker/ColorPickerTypes";
 import { FieldDefDate } from "@root/forms/FormFieldDate/DateField";
 import { FieldDefDropdown } from "@root/forms/FormFieldDropdownSingleSelection";
