+++ src/components/Form/Col/Col.tsx
@@ -3,8 +3,8 @@ import { ElementType, memo, useCallback, useMemo } from "react";
 import { formActions } from "../formActions";
 
 import FormFieldText from "@root/forms/FormFieldText";
-import FormFieldCheckbox from "@root/forms/FormFieldCheckbox";
-import FormFieldChipSingleSelect from "@root/forms/FormFieldChipSingleSelect";
+import FormFieldCheckbox from "@root/components/Field/FormFieldCheckbox";
+import FormFieldChipSingleSelect from "@root/components/Field/FormFieldChipSingleSelect";
 import FormFieldDropdownSingleSelection from "@root/forms/FormFieldDropdownSingleSelection";
 import FormFieldPhoneSelectionDropdown from "@root/forms/FormFieldPhoneSelectionDropdown";
 import FormFieldRadio from "@root/forms/FormFieldRadio";
